Update the references to the old repository owner name
This commit makes updates to the various references to `AnthonyIacono` to replace with the newly formed GitHub organization `canitmux`. I created this organization to ensure that I was able to properly set up a service account with write permissions for the sake of publishing the GitHub Pages site in the pipeline.