Add laptop issue template: Correct syntax errors
Remove the parenthesis from the IDs inside the add laptop issue template form, since this is a syntax error. Similarly, wrap the boolean values with single quotes since this is against the syntax rules.

For more information, see: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/common-validation-errors-when-creating-issue-forms#bodyi-id-can-only-contain-numbers-letters---_ and https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/common-validation-errors-when-creating-issue-forms#bodyi-options-must-not-include-booleans-please-wrap-values-such-as-yes-and-true-in-quotes